requirements.txt
PyPDF2
fastapi
uvicorn
faiss-cpu
sentence-transformers
pyyaml
mlflow
boto3
nltk